// Platform.FoodMe.Admin/Platform.FoodMe.Admin/src/app/routes/ReferentialData/Models/Promotion/OrderStatus.ts
export enum OrderStatus{
    Initial = 0,
    Open = 1,
    Inkitchen = 2,
    PendingPayment = 3,
    BUMPED = 4,
    Ready = 8,
    ASSIGNED = 16,
    ENROUTE = 32,
    Received = 59,
    FindingDriver = 60,
    DriverAccept = 61,
    Indelivery = 63,
    Delivered = 64,
    Arrived = 65,
    Suspended = 96,
    FUTURE = 100,
    Closed = 128,
    FAILURE = 256,
    Canceled = 512,
    UNKNOWN = 1024,
    ForceClosed = 2048,
    REQUESTFORCANCELORDER = 4096,
    ForceCancel = 8192,
    ReadyToPickup = 2022,
}